Start cleaning some ugliness of UMTCli
- seperated commands into seperate file
- A lot of documentation and comments
- seperated all the stuff that gets inherited into seperate file
- remove some unneeded / unused stuff
- added verbose option for new command
- reordered some methods internally
- add #regions for important groups
- add todos for stuff that still needs to be done.
