+++ UndertaleModCli/UndertaleModCli.csproj
@@ -7,6 +7,7 @@
     <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
 	<Nullable>annotations</Nullable>
 	<RollForward>LatestMajor</RollForward>
+	<PackageVersion>1.0.0</PackageVersion>
   </PropertyGroup>
 
   <ItemGroup>
